auth/auth/auth_microservice/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
psycopg2-binary